% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulls_data.R
\docType{data}
\name{gulls_data}
\alias{gulls_data}
\title{This data set contains information about gulls and Toxoplasmosis infections.}
\format{
A data frame with 29 rows and 10 columns:
\describe{
\item{BirdID}{character combination associated with an individual bird}
\item{CWS_band}{number combination associated with an individual leg band}
\item{Year}{Year of capture}
\item{Month}{Month of capture}
\item{Day}{Day of capture}
\item{Toxoplasmosis}{Whether or not bird has the disease Toxoplasmosis. (1=Yes,0=No)}
\item{Bird_weight_kg}{Bird weight at time of capture in killograms}
\item{Head_length_mm}{Bird head length at time of capture in millimeters}
\item{Bill_depth_mm}{Bird bill depth at time of capture in millimeters}
\item{Bill_length_mm}{Bird bill length at time of capture in millimeters}
}
}
\usage{
data(gulls_data)
}
\description{
This data set contains information about gulls and Toxoplasmosis infections.
}
\keyword{data}
\keyword{sets}
